+++ pipe/connection.py
@@ -1,20 +1,8 @@
-from typing import Callable
-
-from common.client.market_socket.async_socket_client import BaseSettingWebsocket
-
-
-class MarketsCoinTickerPriceWebsocket(BaseSettingWebsocket):
-    """티커 웹소켓"""
-
-    def get_websocket_method(self, api: Callable) -> Callable:
-        return api.price_present_websocket
-
-
-class MarketsCoinOrderBookWebsocket(BaseSettingWebsocket):
-    """오더북 웹소켓"""
-
-    def get_websocket_method(self, api: Callable) -> Callable:
-        return api.orderbook_present_websocket
+from common.client.market_socket.async_socket_client import (
+    MarketsCoinTickerPriceWebsocket,
+    MarketsCoinOrderBookWebsocket,
+)
+from config.yml_param_load import SocketMarketLoader
 
 
 class CoinPresentPriceWebsocket(MarketsCoinTickerPriceWebsocket):
@@ -24,7 +12,8 @@ class CoinPresentPriceWebsocket(MarketsCoinTickerPriceWebsocket):
         location: str,
         market: str = "all",
     ) -> None:
-        super().__init__(location, symbol, market)
+        self.market_env = SocketMarketLoader(location=location).process_market_info()
+        super().__init__(symbol=symbol, market=market, market_env=self.market_env)
 
 
 class CoinOrderBookWebsocket(MarketsCoinOrderBookWebsocket):
@@ -34,4 +23,5 @@ class CoinOrderBookWebsocket(MarketsCoinOrderBookWebsocket):
         location: str,
         market: str = "all",
     ) -> None:
-        super().__init__(location, symbol, market)
+        self.market_env = SocketMarketLoader(location=location).process_market_info()
+        super().__init__(symbol=symbol, market=market, market_env=self.market_env)
